requirements (2).txt
### Requirements
librosa
torch==1.13.1
torchaudio==0.13.1
soundfile
scipy
matplotlib
tqdm
numpy
pandas
speechbrain==0.5.15
tensorflow
tensorflow-probability
torchmetrics==0.10.0
seaborn
ipykernel
scaper
thop==0.1.1.post2209072238
openl3
youtube_dl
transformers
bs4
pyroomacoustics==0.7.3
python-sofa==0.2.0
onnx
onnxruntime
torch_tb_profiler
ffmpegio
noisereduce==3.0.0


netCDF4==1.5.7
PyYAML
